refactor(automation): remove legacy subprocess-per-phase loop from loop_runner
The epic #1809 pipeline cutover (#1818) made the queue-based pipeline the
default automation-loop path; the legacy subprocess-per-phase path was kept
byte-for-byte only as the rollback story. Delete it now that the pipeline is
proven in production.

loop_runner.py is now a thin wrapper: CLI parsing, org/repo scope resolution,
PipelineConfig construction, and the token-preflight + run_pipeline dispatch.
Removed: run_phase, _build_phase_argv, _resolve_phase_bin, _phase_env,
_PHASE_FLAGS, process_repo/_process_repo_inner/_process_one_issue,
_run_post_loop_stages, _run_terminal_deferred_issues, _open_peer_pr_claims,
run_loop, PhaseResult/RepoResult, _summarize_loop/_summarize_post_loop,
ALL_POST_LOOP_STAGES post-loop split vestiges, the work-report helpers, the
rate-budget sleep guard, cooperative-shutdown signal handling, and the
--legacy-loop / --pipeline flags plus the HEPH_PIPELINE env escape hatch.
The module docstring is rewritten to describe the thin-wrapper design.

Test re-homing (no coverage lost; 85.66% >= 83%):
- _review_utils.write_work_report / work_report_context coverage moved to
  test_review_utils.py (still used by planner/plan_reviewer).
- PlanReviewer.already_reviewed + planner/reviewer work-count coverage moved to
  test_plan_reviewer.py.
- _count_failing_prs (lives in loop_repo_manager) coverage moved to
  test_loop_repo_manager.py.
- Deleted test_loop_runner_post_loop.py, test_loop_runner_early_exit.py,
  test_loop_runner_file_overlap.py, test_pipeline_shims.py, and
  test_drive_green_pr_discovery.py (pure legacy; product requirements already
  covered under tests/unit/automation/pipeline/).
- Slimmed test_loop_runner.py to parser/scope/dispatch cases; rewrote
  test_pipeline_flag.py for the single pipeline path; dropped the
  --pipeline/--legacy-loop parser specs from test_automation_parsers.py.

loop_runner stays in the coverage omit list (live-CLI entry point).

Closes #1819
